backend/requirements.txt
flask
flask-cors
python-dotenv
pytz
supabase
Jinja2
stripe
requests
schedule